Move the version command out to pkg/version/cmd for reusability
