--- cmd/ignite/cmd/version.go
@@ -1,23 +1,0 @@
-package cmd
-
-import (
-	"io"
-
-	"github.com/spf13/cobra"
-	"github.com/weaveworks/ignite/cmd/ignite/cmd/cmdutil"
-	"github.com/weaveworks/ignite/cmd/ignite/run"
-)
-
-// NewCmdVersion provides the version information of ignite
-func NewCmdVersion(out io.Writer) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "version",
-		Short: "Print the version of ignite",
-		Run: func(cmd *cobra.Command, args []string) {
-			cmdutil.CheckErr(run.Version(out, cmd))
-		},
-	}
-
-	cmd.Flags().StringP("output", "o", "", "Output format; available options are 'yaml', 'json' and 'short'")
-	return cmd
-}
